feat: generate man pages and completions
using mango-cobra
