--- cmd/soft/main.go
@@ -1,110 +1,0 @@
-package main
-
-import (
-	"context"
-	"fmt"
-	"log"
-	"os"
-	"os/signal"
-	"runtime/debug"
-	"syscall"
-	"time"
-
-	"github.com/charmbracelet/soft-serve/config"
-	"github.com/charmbracelet/soft-serve/server"
-	mcobra "github.com/muesli/mango-cobra"
-	"github.com/muesli/roff"
-	"github.com/spf13/cobra"
-)
-
-var (
-	// Version contains the application version number. It's set via ldflags
-	// when building.
-	Version = ""
-
-	// CommitSHA contains the SHA of the commit that this application was built
-	// against. It's set via ldflags when building.
-	CommitSHA = ""
-
-	rootCmd = &cobra.Command{
-		Use:   "soft",
-		Short: "A self-hostable Git server for the command line",
-		Long:  "Soft Serve is a self-hostable Git server for the command line.",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			return cmd.Help()
-		},
-	}
-
-	serveCmd = &cobra.Command{
-		Use:   "serve",
-		Short: "Start the server",
-		Long:  "Start the server",
-		Args:  cobra.NoArgs,
-		RunE: func(cmd *cobra.Command, args []string) error {
-			cfg := config.DefaultConfig()
-			s := server.NewServer(cfg)
-
-			done := make(chan os.Signal, 1)
-			signal.Notify(done, os.Interrupt, syscall.SIGINT, syscall.SIGTERM)
-
-			log.Printf("Starting SSH server on %s:%d", cfg.BindAddr, cfg.Port)
-			go func() {
-				if err := s.Start(); err != nil {
-					log.Fatalln(err)
-				}
-			}()
-
-			<-done
-
-			log.Printf("Stopping SSH server on %s:%d", cfg.BindAddr, cfg.Port)
-			ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-			defer func() { cancel() }()
-			return s.Shutdown(ctx)
-		},
-	}
-
-	manCmd = &cobra.Command{
-		Use:    "man",
-		Short:  "Generate man pages",
-		Args:   cobra.NoArgs,
-		Hidden: true,
-		RunE: func(cmd *cobra.Command, args []string) error {
-			manPage, err := mcobra.NewManPage(1, rootCmd) //.
-			if err != nil {
-				return err
-			}
-
-			manPage = manPage.WithSection("Copyright", "(C) 2021-2022 Charmbracelet, Inc.\n"+
-				"Released under MIT license.")
-			fmt.Println(manPage.Build(roff.NewDocument()))
-			return nil
-		},
-	}
-)
-
-func init() {
-	rootCmd.AddCommand(
-		serveCmd,
-		manCmd,
-	)
-	rootCmd.CompletionOptions.HiddenDefaultCmd = true
-
-	if len(CommitSHA) >= 7 {
-		vt := rootCmd.VersionTemplate()
-		rootCmd.SetVersionTemplate(vt[:len(vt)-1] + " (" + CommitSHA[0:7] + ")\n")
-	}
-	if Version == "" {
-		if info, ok := debug.ReadBuildInfo(); ok && info.Main.Sum != "" {
-			Version = info.Main.Version
-		} else {
-			Version = "unknown (built from source)"
-		}
-	}
-	rootCmd.Version = Version
-}
-
-func main() {
-	if err := rootCmd.Execute(); err != nil {
-		log.Fatalln(err)
-	}
-}
